backend/requirements.txt
# =============================================================================
# requirements.txt — Python Dependencies
# =============================================================================
# Install all with: pip install -r requirements.txt
#
# CONCEPT: Pin major.minor versions (not patch) for reproducible installs
# while still getting security patches (e.g., fastapi>=0.111,<0.112)
# =============================================================================

# Web framework
fastapi>=0.111.0

# ASGI server — runs FastAPI
uvicorn[standard]>=0.30.0     # [standard] includes uvloop (faster event loop) + httptools

# Database
sqlalchemy>=2.0.0              # ORM — async support requires 2.x
asyncpg>=0.29.0                # async PostgreSQL driver (used by SQLAlchemy)
psycopg2-binary>=2.9.0         # sync driver (used by Alembic for migrations)

# Migrations (production alternative to create_all)
alembic>=1.13.0

# Validation
pydantic[email]>=2.7.0         # [email] adds email-validator for EmailStr

# Authentication
PyJWT>=2.8.0                       # JWT encoding/decoding
passlib[bcrypt]>=1.7.4             # password hashing with bcrypt

# Environment variables
python-dotenv>=1.0.0               # loads .env file into os.environ

# HTTP client (for seeding from JSONPlaceholder)
httpx>=0.27.0                      # async HTTP client
